# api/openapi.yaml
openapi: 3.1.0
info:
  title: kafkito REST API
  version: "1"
  summary: Read/write access to Kafka clusters managed by kafkito.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: |
    JSON surface that backs the kafkito web UI. It is primarily a UI
    contract, but stable enough for scripting with `curl`.

    This document is the single source of truth for the HTTP contract
    (ADR-0005). The frontend TypeScript types are generated from it and a Go
    test asserts that every registered route is described here.

    Authentication: every `/api/v1/*` request is validated by kafkito's auth
    middleware. Depending on `KAFKITO_AUTH_MODE` it expects an
    `Authorization: Bearer <JWT>` header (`mock`, `oidc`, `xsuaa` in
    `-tags btp` builds) and answers 401 when it is missing or invalid, or injects a
    synthetic principal (`off`, `-tags devauth` builds only). kafkito has no
    login flow of its own:
    interactive login is handled by an upstream auth proxy (for example the
    SAP approuter or oauth2-proxy), which forwards the bearer token. The validated principal
    feeds the optional RBAC policy. `/healthz` and `/readyz` are never
    authenticated.

    Private (browser-stored) clusters: use the cluster path segment
    `__private__` and send the cluster definition base64-encoded in the
    `X-Kafkito-Cluster` header. Operators can disable private clusters or
    restrict them to a role with `private_clusters.mode`
...

    Conventions:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/clusters:
    get:
      operationId: listClusters
      tags: [clusters]
      summary: List configured clusters with reachability, capabilities and metrics.
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListClustersResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/clusters/{cluster}/topics/{topic}/copy:
    parameters:
      - $ref: "#/components/parameters/Cluster"
      - $ref: "#/components/parameters/Topic"
      - $ref: "#/components/parameters/PrivateClusterHeader"
    post:
      operationId: copyMessages
      tags: [messages]
      summary: Bulk-copy messages into another topic (SSE progress stream).
      description: |
        Server-side copy: reads records from the source topic — the
        `{cluster}`/`{topic}` in the URL — and re-produces them into
        `dest_topic` on the destination cluster named in the request body.
        The destination topic must already exist; kafkito never creates it.

        Exactly one destination cluster field must be set: `dest_cluster` for a
        server-configured cluster, or `dest_cluster_config` for an ad-hoc
        ("private") cluster passed inline — the same payload the
        `X-Kafkito-Cluster` header carries.

        The response is **not** JSON: it is a `text/event-stream` of
        `CopyProgressEvent` objects, because a copy can run far longer than a
        normal request.

        When `to_ts_ms` is omitted the server substitutes the job's start time,
        so copying a live topic terminates instead of tailing it forever:
        records produced after the copy started are not included.

        Records that cannot be reproduced byte-for-byte are left out and
        counted in `skipped` instead of being copied approximately:
        Schema-Registry-decoded payloads (`avro` / `json_schema` /
        `protobuf`), for which only the decoded JSON rendering is available and
        the original wire-format bytes are gone, and masked records, whose
        value, key or header values the source cluster's `data_masking` rules
        replaced with a redacted rendering.

        Copied records carry the same provenance headers the produce endpoint
        injects — `X-Kafkito-Source: true` and, when an identity is available,
        `X-Kafkito-User: <subject>` of the caller. Every source header whose key
        starts with `X-Kafkito-` (matched case-insensitively) is dropped, so the
        original producer's provenance is never carried over.

        Authorization: the source needs `topic:consume` (checked by the RBAC
        middleware from the URL) and the destination needs `topic:produce`
        (checked by the handler against the cluster/topic named in the body).
        A `dest_cluster_config` destination is a private cluster:
        `private_clusters.mode` applies to it (403 `private_clusters_disabled`
        or `private_clusters_forbidden`, checked before the handler validates
        or contacts the destination). Once allowed, RBAC does not apply to the
        destination; the caller supplies their own broker credentials and
        that broker's ACLs apply. Operators can disable private clusters or
        restrict them to a role to limit such copies.

        The copy is neither transactional nor resumable: on error the records
        copied so far stay in the destination topic, and re-running the copy
        duplicates them.
      parameters:
        - $ref: "#/components/parameters/ProdConfirmHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CopyRequest" }
      responses:
        "200":
          description: |
            Copy job started; progress is streamed as Server-Sent Events. A
            failure *during* the copy also arrives here with status 200 — the
            SSE headers are already sent — as a final event carrying `error`.
            Clients must inspect the events, not just the status code.

            Wire format: one `data: <CopyProgressEvent JSON>` line per event,
            separated by a blank line, e.g.
            `data: {"copied":500,"skipped":3}`.
          headers:
            X-Request-Id: { $ref: "#/components/headers/RequestId" }
          content:
            text/event-stream:
              schema: { $ref: "#/components/schemas/CopyProgressEvent" }
              example: { copied: 812, skipped: 5, done: true }
        "400":
          description: >
            Invalid body, missing `dest_topic`, both or neither destination
            field set, destination equal to the source cluster+topic (the copy
            would never terminate), unknown `dest_cluster`, `dest_topic` does
            not exist on the destination (the destination is never
            auto-created), or `preserve_partition` with too few partitions on
            the destination topic.
          headers:
            X-Request-Id: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "428": { $ref: "#/components/responses/ProdConfirmationRequired" }
        "429":
          description: >
            Too many concurrent copy jobs server-wide (limit 4). The body
            carries `code: copy_concurrency_limit`. Copies are long-running and
            each holds broker connections, so the server sheds load rather than
            queueing.
          headers:
            X-Request-Id: { $ref: "#/components/headers/RequestId" }
            Retry-After:
              description: Seconds to wait before retrying.
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502": { $ref: "#/components/responses/BadGateway" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT validated by kafkito (auth modes `mock`, `oidc` and `xsuaa`). In mode
        `off` (`-tags devauth` builds only) no token is required and a
        synthetic principal is used.
  headers:
    RequestId:
      description: >
        Request correlation id, set on every response. Reuses the inbound
        `X-Vcap-Request-Id`, the trace-id of a W3C `traceparent` or an inbound
        `X-Request-Id` when present (in that order), otherwise freshly
        generated. Matches the `request_id` field in the server logs.
      required: true
      schema: { type: string }
  parameters:
    Cluster:
      in: path
      name: cluster
      required: true
      schema: { type: string }
      description: >
        Cluster name from the kafkito config, or `__private__` together with
        the `X-Kafkito-Cluster` header for a browser-stored cluster.
    Topic:
      in: path
      name: topic
      required: true
      schema: { type: string }
...
    PrivateClusterHeader:
      in: header
      name: X-Kafkito-Cluster
      required: false
      schema: { type: string, contentEncoding: base64 }
      description: >
        Base64-encoded JSON `ClusterConfig` of a private (browser-stored)
        cluster, max 8 KiB decoded. Honoured when the `{cluster}` path segment
        is `__private__` (required then). A malformed header is rejected with
        400. When `private_clusters.mode` does not allow the caller private
        clusters, the header is never read: a `__private__` request gets 403
        (`private_clusters_disabled` or `private_clusters_forbidden`), and any
        other request ignores it. On an allowed request kafkito's RBAC does
        not apply to the private cluster; the broker's ACLs do.
    ProdConfirmHeader:
      in: header
      name: X-Kafkito-Confirm-Prod
      required: false
      schema: { type: string, enum: ["true"] }
      description: >
        Must be `true` when the target cluster is marked `is_prod`; the request
        is rejected with 428 otherwise. Send only after explicit user
        confirmation.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
        WWW-Authenticate:
          description: Always `Bearer realm="kafkito"`.
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: >
        RBAC denied the action (`error: forbidden` plus `resource` and
        `action`, or `code: rbac_denied`), the broker rejected the
        credential's ACLs (`code: kafka_not_authorized`), or
        `private_clusters.mode` does not allow the caller private clusters
        (`code: private_clusters_disabled` for mode `off`,
        `private_clusters_forbidden` for mode `role` without
        `private_cluster:use`).
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ProdConfirmationRequired:
      description: >
        Cluster is marked `is_prod` and the `X-Kafkito-Confirm-Prod: true`
        header is missing (`code: production_confirmation_required`).
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadGateway:
      description: >
        Upstream Kafka / Schema Registry error. Details are logged
        server-side only (`code: kafka_upstream`). When the outbound address
        guard refused the connection because a broker or Schema Registry
        address of a private cluster resolves to a blocked range (loopback,
        link-local, multicast, unspecified), the code is
        `private_cluster_address_blocked` and the message is static; it never
        names the address. Test connection reports the affected brokers.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      description: Error body emitted by all JSON error paths.
      properties:
        error:
          type: string
          description: Human-readable message (`unauthorized`, `forbidden`, `not found`, ...).
        code:
          type: string
          description: >
            Machine-readable code where one exists, e.g. `kafka_upstream`,
            `private_cluster_address_blocked`, `private_clusters_disabled`,
            `private_clusters_forbidden`,
            `production_confirmation_required`, `copy_concurrency_limit`,
            `kafka_message_too_large`, `kafka_not_authorized`, `rbac_denied`,
            `topic_consumers_timeout`, `value_masked`, `invalid_request` (the
            request does not match this document; `error` names the parameter
            or body field and the violated rule, never the submitted value).
        message:
          type: string
          description: Detail for 401 responses from the auth middleware.
        resource:
          type: string
          description: RBAC-denied resource as `<type>:<name>` (403 only).
        action:
          type: string
          description: RBAC-denied action (403 only).
...
    ListClustersResponse:
      type: object
      required: [clusters]
      properties:
        clusters:
          type: array
          items: { $ref: "#/components/schemas/ClusterInfo" }
    ClusterInfo:
      type: object
      required: [name, reachable, is_prod, auth_type, tls, schema_registry]
      properties:
        name: { type: string }
        reachable: { type: boolean }
        error: { type: string }
        is_prod: { type: boolean }
        auth_type: { type: string }
        tls: { type: boolean }
        schema_registry: { type: boolean }
        capabilities: { $ref: "#/components/schemas/Capabilities" }
        broker_issues:
          type: array
          description: >
            Test connection only: advertised brokers that could not be
            reached although a seed broker answered. Absent when every
            broker answered; a non-empty list implies `reachable: false`.
          items: { $ref: "#/components/schemas/BrokerIssue" }
        brokers_skipped:
          type: integer
          description: >
            Test connection only: how many advertised brokers were not
            checked because the probe dials at most 64 of them. Absent when
            every advertised broker was checked. On its own it does not make
            `reachable` false.
        brokers:
          type: integer
          description: Aggregate metrics below are best-effort and absent until known.
        topics: { type: integer }
        groups: { type: integer }
        total_messages: { type: integer, format: int64 }
        total_lag: { type: integer, format: int64 }
        total_rate_per_sec: { type: number, format: double }
    BrokerIssue:
      type: object
      required: [node_id, host, port, reason, error]
      properties:
        node_id: { type: integer, format: int32 }
        host:
          type: string
          description: Host the broker advertises in its metadata response.
        port: { type: integer, format: int32 }
        reason:
          type: string
          enum: [blocked, unreachable]
          description: >
            `blocked`: the address resolves to a range the outbound address
            guard refuses for private clusters. `unreachable`: the address
            is allowed but the broker did not answer (refused, timeout, TLS
            or SASL failure).
        error:
          type: string
          description: Raw connection error, for diagnosis.
    Capabilities:
      type: object
      required: [describe_cluster, list_topics, describe_configs, list_groups, create_topic, delete_topic, alter_configs, probed_at]
      properties:
        describe_cluster: { type: boolean }
        list_topics: { type: boolean }
        describe_configs: { type: boolean }
        list_groups: { type: boolean }
        create_topic: { type: boolean }
        delete_topic: { type: boolean }
        alter_configs: { type: boolean }
        errors:
          type: object
          additionalProperties: { type: string }
        probed_at: { type: string, format: date-time }
...
    CopyRequest:
      type: object
      required: [dest_topic]
      additionalProperties: false
      description: >
        Body of the topic copy endpoint. Exactly one of `dest_cluster` /
        `dest_cluster_config` must be set.
      properties:
        dest_cluster:
          type: string
          description: >
            Name of a server-configured destination cluster. Mutually exclusive
            with `dest_cluster_config`.
        dest_cluster_config:
          $ref: "#/components/schemas/ClusterConfig"
          description: >
            Ad-hoc ("private") destination cluster, same shape the
            `X-Kafkito-Cluster` header carries. Mutually exclusive with
            `dest_cluster`. Subject to `private_clusters.mode`; once allowed,
            RBAC does not apply to the destination.
        dest_topic:
          type: string
          description: Destination topic. Must already exist.
        partition:
          type: integer
          format: int32
          description: Single source partition. Absent = all partitions.
        from_ts_ms:
          type: integer
          format: int64
          description: Inclusive lower bound on source record timestamps.
        to_ts_ms:
          type: integer
          format: int64
          description: >
            Exclusive upper bound on source record timestamps. When omitted the
            server substitutes the job's start time, so a copy of a live topic
            terminates instead of tailing it forever.
        limit:
          type: integer
          format: int64
          description: Maximum number of records to copy. Absent = no limit.
        preserve_partition:
          type: boolean
          description: >
            Produce each record to the same partition number it came from.
            Requires the destination topic to have at least as many partitions
            as the highest source partition, otherwise the request is rejected
            with 400.
    CopyProgressEvent:
      type: object
      required: [copied]
      description: >
        One SSE event of the copy stream, sent as a `data: {json}` line pair.
        Progress events arrive periodically — one right after the stream opens
        and at least one per fetched page — and the final event has
        `done: true`.
      properties:
        copied:
          type: integer
          format: int64
          description: Records produced to the destination topic so far.
        skipped:
          type: integer
          format: int64
          description: >
            Source records deliberately left out because they cannot be
            reproduced byte-for-byte (Schema-Registry-decoded, or masked
            values, keys or header values). Omitted while 0.
        done:
          type: boolean
          description: True on the final event only; omitted otherwise.
        error:
          type: string
          description: >
            Set on the final event when the job aborted. Omitted when empty.
            Errors after the stream opened still carry HTTP status 200.
    ClusterConfig:
      type: object
      required: [brokers]
      description: >
        Ad-hoc ("private") cluster definition supplied per request instead of
        from the server config — the same JSON the `X-Kafkito-Cluster` header
        carries (base64-encoded there). Broker and Schema Registry hosts are
        subject to the server's outbound-host (SSRF) policy.
      properties:
        name: { type: string }
        is_prod: { type: boolean }
        brokers:
          type: array
          minItems: 1
          items: { type: string }
        auth:
          type: object
          properties:
            type:
              type: string
              enum: ["", none, plain, scram-sha-256, scram-sha-512]
              description: >
                Canonical lowercase values; empty means `none`. `plain` /
                `scram-*` require `username` and `password`. The
                `X-Kafkito-Cluster` header is not schema-validated (the server
                decodes it) and still accepts these values case-insensitively
                and trimmed, for compatibility with stored private clusters.
            username: { type: string }
            password: { type: string, format: password }
        tls:
          type: object
          properties:
            enabled: { type: boolean }
            insecure_skip_verify: { type: boolean }
        schema_registry:
          type: object
          properties:
            url: { type: string }
            username: { type: string }
            password: { type: string, format: password }
            insecure_skip_verify: { type: boolean }
        data_masking:
          type: [array, "null"]
          items: { $ref: "#/components/schemas/MaskingRule" }
    MaskingRule:
      type: object
      properties:
        topics: { type: array, items: { type: string } }
        fields:
          type: array
          items: { type: string }
          description: JSONPath expressions whose values are replaced.
        regex:
          type: array
          items:
            type: object
            properties:
              match: { type: string }
              replacement: { type: string }
        replacement: { type: string, description: 'Replacement for `fields`; empty = "***".' }
        targets:
          type: array
          items: { type: string, enum: [value, key, headers] }
          description: Record parts the rule masks. Absent or empty = `[value]`.
        headers:
          type: array
          items: { type: string }
          description: >
            Go regex patterns on header keys that restrict the `headers` target
            to matching headers. Absent or empty = every header.
...